+++ Interface/bin/Debug/Interface.vshost.exe.manifest
@@ -78,14 +78,14 @@
     </dependentAssembly>
   </dependency>
   <dependency>
-    <dependentAssembly dependencyType="install" allowDelayedBinding="true" codebase="Interface.exe" size="1040896">
+    <dependentAssembly dependencyType="install" allowDelayedBinding="true" codebase="Interface.exe" size="1041408">
       <assemblyIdentity name="Interface" version="1.0.0.0" language="neutral" processorArchitecture="x86" />
       <hash>
         <dsig:Transforms>
           <dsig:Transform Algorithm="urn:schemas-microsoft-com:HashTransforms.Identity" />
         </dsig:Transforms>
         <dsig:DigestMethod Algorithm="http://www.w3.org/2000/09/xmldsig#sha256" />
-        <dsig:DigestValue>Rhvvan4cLA+ZWJHZw8J+/eXc9Z8/4sh4JOUyBKdPohk=</dsig:DigestValue>
+        <dsig:DigestValue>qGDbkBg+SfrX91pIU+hIYA2CCm2BdncjfZjEL8btToA=</dsig:DigestValue>
       </hash>
     </dependentAssembly>
   </dependency>
